test(frontend): split diagnostics CLI model coverage
